# Optimise DependencyContainer lookups