--- a/osu.Framework/Allocation/CacheInfo.cs
+++ b/osu.Framework/Allocation/CacheInfo.cs
@@ -6,7 +6,7 @@ using osu.Framework.Extensions.TypeExtensions;
 
 namespace osu.Framework.Allocation
 {
-    public readonly struct CacheInfo
+    public readonly struct CacheInfo : IEquatable<CacheInfo>
     {
         /// <summary>
         /// The name of the cached member.
@@ -38,5 +38,20 @@ namespace osu.Framework.Allocation
         internal CacheInfo WithType(Type type) => new CacheInfo(type, Name, Parent);
 
         public override string ToString() => $"{nameof(Type)} = {Type?.ReadableName()}, {nameof(Name)} = {Name}, {nameof(Parent)} = {Parent?.ReadableName()}";
+
+        public override bool Equals(object obj) => obj is CacheInfo cacheInfo && Equals(cacheInfo);
+
+        public bool Equals(CacheInfo other) => Name == other.Name && Parent == other.Parent && Type == other.Type;
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                var hashCode = (Name != null ? Name.GetHashCode() : 0);
+                hashCode = (hashCode * 397) ^ (Parent != null ? Parent.GetHashCode() : 0);
+                hashCode = (hashCode * 397) ^ (Type != null ? Type.GetHashCode() : 0);
+                return hashCode;
+            }
+        }
     }
 }