I noticed the memory usage of this (was somewhere in the top 5 on world.execute(me)), but turns out to have a significant impact on time as well:

```diff
  | Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
  |------- |---------:|---------:|---------:|------:|------:|------:|----------:|
- |    Get | 21.34 us | 1.099 us | 3.136 us |     - |     - |     - |     936 B |
+ |    Get | 870.2 ns | 48.47 ns | 134.3 ns |     - |     - |     - |         - |
```

(per `.Get()` call)